added code to sensor component
