+++ package.json
@@ -9,10 +9,13 @@
   },
   "dependencies": {
     "-": "^0.0.1",
+    "@expo-google-fonts/lato": "^0.2.2",
     "@react-navigation/native": "^6.0.10",
     "braces": "^3.0.2",
     "expo": "~43.0.0",
+    "expo-app-loading": "~1.2.1",
     "expo-cli": "^5.4.2",
+    "expo-font": "~10.0.3",
     "expo-sensors": "~11.0.3",
     "expo-status-bar": "~1.1.0",
     "mem": "^9.0.1",
@@ -22,6 +25,7 @@
     "react-native": "0.64.3",
     "react-native-accurate-step-counter": "^2.0.4",
     "react-native-shake": "^5.1.1",
+    "react-native-vertical-text": "^1.0.5",
     "react-native-web": "0.17.1",
     "styled-components": "^5.3.3"
   },
